+++ src/index.tsx
@@ -1,6 +1,7 @@
 import "virtual:windi.css";
 export { createSelect } from "./create-select";
 export { createOptions } from "./create-options";
+export { fuzzySort, fuzzySearch, fuzzyHighlight } from "./fuzzy";
 export {
   Select,
   Container,